Fixed Task index
If the player had 2 scrap items in their inventory and entered the ship without having them selected then when the user would press G to drop, it would start from the first available scrap index and delete that scrap item instead of the one currently being held.

This fixes that issue so that the item currently held will be removed.